requirements.txt
streamlit>=1.32.0
pdfplumber>=0.11.0
pytesseract>=0.3.13
Pillow>=10.0.0
requests>=2.31.0
groq>=0.5.0
pandas>=2.0.0
python-docx>=1.1.0
fpdf2>=2.7.0
